Добавлено оздание строки "Message"
